FIX(tsconfig): change absolute paths and imports
